basic schema converter
